intuitiveobjects-rag-api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
motor==3.7.0
oauthlib==3.2.2
proto-plus==1.26.1
protobuf>=3.20.2,<5.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.1
pyparsing==3.2.3
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.0
accelerate==0.34.2
aiofiles==0.6.0
#aiohttp==3.7.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.1
attrs==23.2.0
av==13.0.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitarray==2.9.2
bitsandbytes==0.44.0
blinker==1.8.2
build==1.2.1
Byaldi==0.0.4
camelot-py==0.11.0
catalogue==2.0.10
cffi==1.16.0
chardet==3.0.4
chroma-hnswlib==0.7.3
chromadb==0.5.3
ci-info==0.3.0
#clip==1.0
cmd2==1.5.0
colbert-ai==0.2.21
colorama==0.4.6
coloredlogs==15.0.1
colpali_engine==0.3.1
commonmark==0.9.1
configobj==5.0.8
configparser==7.0.0
contourpy==1.3.0
cryptography==42.0.8
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.6.7
datasets==3.0.1
DateTime==5.5
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
ecdsa==0.19.0
einops==0.8.0
et-xmlfile==1.1.0
etelemetry==0.3.1
eval_type_backport==0.2.0
faiss-cpu==1.8.0.post1
fastapi-cli==0.0.4
ffmpy==0.4.0
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
ghostscript==0.7
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-ai-generativelanguage==0.6.6
google-generativeai==0.7.2
gotrue==2.11.0
GPUtil==1.4.0
#gradio_client==1.3.0
greenlet==3.0.3
#groq==0.11.0
grpcio==1.64.1
grpcio-status==1.62.2
httpcore==1.0.6
httptools==0.6.3
httpx
huggingface-hub==0.23.4
humanfriendly==10.0
identify==2.5.36
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
kubernetes==30.1.0
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.16
langchain-google-genai==1.0.7
langchain-text-splitters==0.2.2
langsmith==0.1.85
looseversion==1.3.0
lxml==5.2.2
markdown-it-py==3.0.0
marker-pdf==0.2.16
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
mdurl==0.1.2
ml_dtypes==0.5.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
#mteb==1.6.35
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
ollama==0.3.2
onnxruntime==1.18.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.2
pdfreader==0.1.15
pdftext==0.3.10
peft==0.11.1
pillow==10.4.0
pip==24.3.1
pipdeptree==2.24.0
pluggy==1.5.0
polars==1.8.2
#postgrest-py
posthog==3.5.0
prov==2.0.1
psutil==6.0.0
py==1.11.0
pyarrow==16.1.0
pycparser==2.22
pycryptodome==3.20.0
pydantic-settings==2.3.4
pydeck==0.9.1
pydot==3.0.1
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.10
pymupdf4llm==0.0.17
PyMuPDFb==1.24.10
pypdf==5.0.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pytesseract==0.3.10
pytest==6.2.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-jose==3.3.0
python-magic==0.4.27
pytrec-eval-terrier==0.5.6
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
qwen-vl-utils==0.0.8
rank-bm25==0.2.2
rapidfuzz==3.9.4
rdflib==6.3.2
#realtime-py==0.1.3
referencing==0.35.1
regex==2024.5.15
rfc3986==1.5.0
#rich==10.2.0
rpds-py==0.19.0
ruamel.yaml==0.17.17
ruff==0.6.8
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.1.0
setuptools==70.1.1
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.31
srsly==2.4.8
#streamlit==1.36.0
supabase==2.13.0
surya-ocr==0.4.15
sympy==1.12.1
tabula-py==2.9.3
tabulate==0.9.0
tenacity==8.4.2
texify==0.1.10
threadpoolctl==3.5.0
tika==2.6.0
tokenizers==0.20.0
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
#torch==2.4.1+cpu
#torchaudio==2.4.1+cpu
#torchvision==0.19.1+cpu
tornado==6.4.1
tqdm==4.66.4
traits==6.3.2
transformers==4.45.1
triton==3.0.0
typer==0.12.3
typing-inspect==0.9.0
tzdata==2024.1
ujson==5.10.0
Unidecode==1.3.8
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
#websockets==9.1
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.19.2
zope.interface==6.4.post2